requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.1
httpx==0.27.0
psycopg2-binary==2.9.9 # For PostgreSQL connection
asyncpg==0.29.0 # For SQLAlchemy async support with PostgreSQL
openai==1.35.10
fastapi-injector==0.1.1
